Consider addresses with same `name` equals
When comparing addresses, in order to be compatible with all
possible combinations, we should not use legacy attributes
such as `firstname` and `lastname` in the comparison, as they
may end up generating false negatives. Instead, using `name`
should be safer. For example:

  Spree::Address.new(name: 'Mary Jane Watson')

should be equal to:

  Spree::Address.new(firstname: 'Mary Jane', lastname: 'Watson')

This should happen only when the application has already started
moving towards the unique `name` field, i.e. when the preference
`Spree::Config.use_combined_first_and_last_name_in_address` is
set to `true`.

Differently, `firstname` and `lastname` should be accounted for
when checking equality, so that:

  Spree::Address.new(firstname: 'Mary Jane',  'Watson')

won't be equal to:

  Spree::Address.new(firstname: 'Mary',  'Jane Watson')