Fix solidus stock locations sorting
The order of the stock locations that is returned from the configured
sorter is lost in the SimpleCoordinator.

Previously, the  #sort_availability
method was used to re-order the stock locations but this was
lost during the refactor to allow extending the default Solidus
allocation logic.

This commit re-orders the stock locations for all methods
called by the stock allocators.