Configure Active Storage on dummy apps for extensions
Copying the .yml configuration file into config/storage/test.yml
will automatically configure Active Storage in that env [1].

The configuration key needs to be the same of the service
set into config/environments/test.rb.

When testing extensions against old versions of Solidus, it
will not use Active Storage so, it should not be harmful to
have it configured anyway.

[1]: https://github.com/rails/rails/blob/66b861f24cd8ca34b726f77b46f6b25b21f0b972/activestorage/lib/active_storage/engine.rb#L115-L132