Remove explicitely running Active Storage migrations in Dummy App
This is not needed, those migrations are already added by default
on any new Rails Application and are not harmful for testing
purposes. We could skip them in the installer with
`--skip-activestorage` and add them only when we are actually using
it but since the long-term goal is to only support Active Storage
we can save some ifs and keep those migrations even if we are not
using the tables generated.