Do not allow prices with nil amount
This logic will help in an upcoming work that will copy all prices from
a master variant over its children variant (thus enforcing that no
"empty" prices are created).

This overtolerance was introduced in 58ec5cd5. That commit introduced
the `Spree::Price` model to allow dealing with multiple currencies. It
allowed `amount` to be `nil` for master variants, variants with no
master variant, and variants with a master variant with no price. AFAIK,
these three scenarios are no considered by our business logic.

In fact, the logic was removed 5 days later in 9d563e3.

58ec5cd5 also allowed for `amount` to be `nil` in any other case, but
then it copied it from the master variant:

https://github.com/solidusio/solidus/commit/9d563e38ee4ac3f6d3f656f204bd1cc84348bb4e#diff-c083f45179b62dc8c47e4fdcbf7ee31da80552100fc0258238817086e46e2925L35

However, this was also done in the controller layer and it's still in
place:

https://github.com/solidusio/solidus/blob/3cc2d3e492b0aa6c84ac66814db26bea4f17d2e0/backend/app/controllers/spree/admin/variants_controller.rb#L19