Copy all prices from master over children variants
This commit changes functionalities in three ways:

- Backend: Before this change, when a user created a new variant the
  form was pre-populated only with a single price field for the default
  currency. In the case the user removed the amount field and submitted
  the form, the created price had an amount equal to 0. Now, all the
  prices created for a given product (for a given master variant, in
  fact) get pre-populated in a nested form. In the case one of the
  amounts is removed by the user, the form becomes invalid on submission
  and forces the user to set an amount for every price.

- API: Before this change, the user was able to create a variant 1)
  manually assigning a single price in the default currency through the
  `price` attribute, and 2) letting it inherit the default price from
  the master variant when `price` attribute was not present. Now, when
  not overridden all master prices get inherited. Otherwise, the user
  can specify both through `price` (for the default currency) and
  `prices_attributes` attributes.

- Product duplicator: Before this change, only the default master price
  was copied. Now, all prices are taken into account.

Internally, the `set_price` callback on `Variant` has been removed in
order to allow that flexibility. The behavior has been moved to a layer
one step closer than where it should actually be: A service layer. It
will help when we tackle this long-term goal. It also means that when we
manually create a variant we must specify the default price for it.

Having three rails associations between `Variant` and `Price` is a
source of inconsistencies, as when we add a `Price` in-memory to one of
them the others don't get updated. It should be changed as follow-up
work. For now, `Variant#any_price?` hack has been added. In the same
way, `Variant#find_or_build_default_price` didn't find it when it still
was in-memory. `Variant#default_price_from_memory` have been added to
fix it.