Configure preferences in Spree::Preferences::Persistable
We can't do this in Spree::Base as of Rails 6.1 as Rails 6.1 expects
there to always be a preferences column if one calls `serialize
:preferences, Hash`.

Only a few classes have that column. This moves the code from
Spree::Base into a module, and includes that in those classes that
actually have a preferences column.

There are no deprecations in this commit, as the deprecation work is
done on the v2.11 branch.