Move currently_valid_prices to a method
Having `currently_valid_prices` as an association between `Variant` and
`Price` made it a source of inconsistencies with the undecorated
`prices` association:

```
Spree::Variant.new.tap do |v|
  v.currently_valid_prices << Spree::Price.new
end.prices.any? # => false
```

We're adding tests to the method, making explicit that at this moment a
price that belongs to a country is prioritized over a generic price when
it comes to choose which one is showed as the default. Probably it's not
desired behavior and we should change it in a follow-up work.

We should also consider renaming it to something more communicative. The
current name conveys the idea that we're filtering some records out,
while in fact we're just ordering them.