Allow accessing preferences on models that do not have any set
We may have models that are supposed to have preferences but are not
defining them explicitly because they are not needed.

For example, when defining a custom calculator that does not need any
preference. Core code expects that preferences still responds with a
Hash instead of nil because that's how it worked before b947015bc28c.

That commit is needed because otherwise Rails would serialize the object
differently on models that do not use preferences, because seralize is
now lazy executed. Example of the wrong serialization without it:

       expected: #<Spree::Address id: 4, firstname: "John", lastname: "Von Doe", address1: "10 Lovely Street", address...00 +0000", updated_at: "2021-03-17 08:09:00.244459000 +0000", name: "John Von Doe", preferences: {}>
            got: #<Spree::Address id: 4, firstname: "John", lastname: "Von Doe", address1: "10 Lovely Street", address...0 +0000", updated_at: "2021-03-17 08:09:00.244459000 +0000", name: "John Von Doe", preferences: nil>

This commit introduces a hack to have both things. When preferences is
empty at database level, it's safe to always return a Hash, because
that's how the data would have been deserialized anyway. This allows
us to call `preferences[:something]` on models that do not explicitly
define any preference.