Add price_for_options and deprecate price_for
The price_for method does not return a Spree::Price as described.
Instead it does some frontend work and returns a Spree::Money, limiting
the usability of this method, as the price record itself can have useful
information on it.

This deprecates the existing method in favor of one that returns a
Spree::Price.