Add custom price_for_options method to variant
We need to safely ensure the new price_for_options method in the price
selector will not break for those that have a custom price selector.

If the new method is defined in the price selector class this will
behave normally, and if their customization does not allow for it, we
will instead return an unpersisted Spree::Price from the Spree::Money
object that the deprecated price_for method returns.

Spree::Product can now delegate price_for_options to master variant