Spec for customer return referencing existing items
This change adds a pending test which demonstrates an issue with the
`create` action when attempting to pass a reference to existing return
items from a previously created return authorization. Rails does not
allow us to update the association on the return items to the customer
return while creating the return.

Co-authored-by: Mike Conlin <mike@super.gd>