Patched FileNotFoundError bug on failed image downloads.
In rare circumstances, an image would fail to upload to a third party
storage service, causing the application to hard crash when trying to
retrieve an image resource. This commit implements a fix to return a
place holder image to alert the user of an error and also prevents the
application from crashing. This fix will not, however, save any errors
associated with the failed upload such as credential errors.
This fix was tested by artificially changing the blob record to force
a FileNotFoundError.

Commit History:
Implemented fix for AS::FileNotFound Error"
Implemented feature test for potential FileNotFoundError
Added conditional to test because it is specific to Active Storage