Bump removal horizon for 3.x deprecations
Now that we are close to the release of 3.0, we need to
change the deprecation horizon of our Spree::Deprecation
class in order to print the right messages to user when
we emit a deprecation warning.

This change will tell users that code deprecated using
Spree::Deprecation will be removed in Solidus 4.0.