Default `belongs_to_required_by_default` in the dummy app
`load_defaults` was added in the dummy app in a recent commit. This one
fixes some tests so that we no longer need to use the old default for
`belongs_to_required_by_default` setting.

As the taxonomy association for taxons is now required, and we have
indeed treated it as required so far, we just need to add it when we
create taxons resources.