Updated decorator pathing for new applications
The current functionality of finding decorators was by pulling files
with the 'decorator' suffix in a file name. This seems redundant
and forces modules to end with "Decorator." It makes more sense
that decorators should be discovered based on their inclusion in the
"app/decorators" sub-folder of the user's application.
Updated guides to reflect name spacing changes.