Add UUID to StoreCredit#generate_authorization_code
The value returned by `StoreCredit#generate_authorization_code`
depends on time, which means that it may not be unique when a
store credit is saved more than once at the same moment. This
is not very likely to happen, but not impossible at all.

By introducing a UUID into the code we can be sure that this
code will end up being unique. The previous format of the code
is retained, since there may be applications relying on it.