Change frontend to allow multiple variants
Before, the front-end _cart_form partial and OrdersController#populate
only allowed for one variant to be sent. This changes that behaviour
to allow for multiple variants, which increases flexibility for
custom front-ends where add-ons can be included on the product page.

The old parameters :variant_id and :quantity are still allowed,
however gives a deprecation warning. The new parameter is:
variants: [{ :variant_id, :quantity }]
The _cart_form partial and other specs that reference #populate
have been updated to use this.