Simplify proposed shipments calculations
We pass the order as a parameter to the estimator and avoid us
depending on the complex AR associations cache. As users can provide
both their own coordinator and estimator classes, we could have
backward compatibility issues both from the caller and the callee. As a
first step, we keep the present behavior and provide the order anyway
as an optional argument.

We prefer the order passed into the intializer over
`package.shipment.order`.

This is undoing a change that was made in 2016
https://github.com/solidusio/solidus/commit/796ed7fa49e793360a1b47b6df03a7bc2a116ccf