Fix admin portugues locale
When changing the admin backend language to Português (PT), it will look for the file select2_locale_pt-PT.js that does not exist.

To Reproduce:
1. Open the admin backend
2. Select Português (PT) in the language selection menu at the bottom right

It is looking for the right file. The file select2_locale_pt-PT.js must be renamed to select2_locale_pt.js and the two references "pt-PT" must be changed to "pt"