Revert "Deprecate callback setting variant price from master"
For now, we've decided to put the greater goal behind this change
(inheriting all prices from master when creating variants) on hold. The
reason for that is that we have higher priorities and that we'll revisit
refactoring core logic when we implement more events or services on
Solidus.

As the reverted commit raised a deprecation warning, it's better to wait
until we have a clearer direction in mind.