Fix Ransack error when searching for orders by customer name
There's an issue in 3.0 where, if you try to search for an
order by the customer's name, Ransack crashes.

The problem was introduced inadvertently in 94bfaa7[1]. This
Ransackable attribute was supposed to be removed completely,
since it is not a Ransack alias anymore, but an actual DB column
that lives on the `bill_address` association.

[1]: https://github.com/solidusio/solidus/commit/94bfaa7a1796cf0dbe45792855e99b2d255af1e5