Remove unused `ShippingRateTaxer` service object
This class was originally used by `Spree::Stock::Estimator` to apply
taxes to a shipping rate, much like `Spree::Tax::OrderAdjuster` does
for orders.

The only call site was then removed in 3639585[1] with the goal of
improving the performance of shipping rate estimation, but the actual
service object was not removed, and is currently sitting unused in the
codebase.

[1]: https://github.com/solidusio/solidus/commit/36395855f7fa6670bd0d63b12cc238b531ea21d9