Make option value to variant association unique
I can't think of a reason why someone would want to associate the same
option value with a given variant multiple times. This makes the index
on the variant/option_value_id columns of the
spree_option_values_variants join table a unique index so we can't have
a variant that has the same option value multiple times.

There's a vaguely related issue over on solidus_importer where the
option value processor attempts to do this if you try to reimport
products. This will cause it to error out instead.