Fix displaying of deleted variants in admin panel
Because of the default scope added in
https://github.com/solidusio/solidus/pull/3488/commits/466251e98255a90fc4c8cc7b975b9e6a7eb11ceb,
the `discarded` scope currently operates on the default `kept` scope and
returns an empty set.

This fixes the scope for checking the `discarded` variants as
suggested here https://github.com/jhawthorn/discard#default-scope, to
correctly display the "deleted variants" filter in the admin variants
page.