Fix gateway_error when no order is defined
If the current controller raises a gateway error, but does not have
`@order` defined, this error handler will raise a NoMethodError when it
tries to access `nil.errors`.

We can resolve this by falling back to an empty Order instance if none
is defined. This maintains the existing behaviour, and as the response
only serializes the errors hash, there shouldn't be any surprises with
null fields, etc.

Signed-off-by: Alex Blackie <alex@alexblackie.com>