Remove the upgrade task and point to additional steps from the update generator
Having both an `upgrade` task and an `update` generator is confusing. On
top of that, it's better to have actionable items, which can be
undesirable in some situations, the most explicit as possible. For this
reason, we leave in place the `update` generator, which is just a safe
path to update preferences, and remove the `upgrade` task. Instead, we
add a generic message into the `update` generator pointing to the
detailed instructions in the Changelog.

In particular, at this point, the `update` task was calling to a single
task introduced in #3987. We still keep the task reachable by end-users,
but we rename it slightly to be more friendly.