Add new scopes for searching products by SKU
This scope is currently only in use on the product search page.
Unfortunately, it did not return the correct results when attempting to
search with "deleted" products by SKU. This is due to the scope
excluding any discarded variants by default.

To remedy the issue, we've addded a second scope that will include the
discarded variants in the results. We've also renamed the old scope to
better reflect what it is doing and have added a deprecation for the old
method name.