Fix searching discarded products in admin by SKU
Now that we've got a new scope we can use that will correctly return
products with discarded variants, we've updated the admin interface to
use it.

This meant adding a new Backbone view to help control the two inputs we
now have on the page, one for each scope. It will hide and disable the
appropriate input based on the value of the "with deleted" checkbox.