Rename Order#ensure_updated_shipments
The naming of this method is confusing and it may induce developers
to believe it refreshes shipments while it actually only destroys
them on a standard Solidus store.

The method definition is thoroughly tested in the codebase, and the
wanted behavior is to actually remove shipments without recreating
them back.

The old name is still available but deprecated with a message that
suggests using the new one.

FYI if the original order's state machine is patched for example
by removing the `address` state, then shipments will be actually
recreated since the order will be `next`ed to the `delivery` state
and a callback will create them again.