Fix loading core on Rails < 6.1
To avoid warning about `Spree.load_defaults` not being defined at a
moment when `bin/rails g solidus:install` has not even been called, the
`Spree::Core` engine is checking whether the `spree.rb` initializer
already exists (as it's there where `#load_defaults` should be called).

We're looking for `spree.rb` within the directories associated with the
`Rails::Paths::Path` instance for initializers. However, before this
commit, we were using `Rails::Paths::Path#paths` method, which is only
available on Rails >= 6.1.0. That made the loading of `Spree::Core`
crash.

As `Rails::Paths::Path` is an `Enumberable`, we can use `any?` directly
from the instance and manually build the `Pathname` for each yielded
string instead.

We also added tests to cover the previous failure and changed the naming
to make the purpose clearer.

Fixes #4178