Fix discarded duplicated products bug
When a product is cloned and then discarded without changing the cloned
variants SKU, if the product is ever cloned again then the cloning
process will fail with a variant and master variant SKU validation
error. This issue occurs because products and their variants are soft
deleted and therefore the copied SKU will not be unique during the
second cloning.

The previous variant SKU uniqueness scope did not work for this specific
case so this conditional was added on top of it.

This commit also includes a validation spec for the variant SKU and a
spec to catch the deleted duplication error.