Introduce a configuration value for migration path
Since Rails 6 multiple database support is a native feature, which
allows users to configure separate read/write databases. This change
allows users to configure the path to the migrations folder from the
default "db/migrate" if they are using the multiple database support and
have specified a different database for the Solidus engine. This will
ensure the check that all migrations are installed can look at a user
configurable folder.

One change required to make the folder path configurable is to run the
migration check after 'load_config_initializers' instead of before. This
will allow the value of `migration_path` configuration to be set in an
initializer.

Co-authored-by: Adam Mueller <adam@super.gd>
Co-authored-by: Alex Blackie <alex@super.gd>
Co-authored-by: Benjamin Willems <benjamin@super.gd>
Co-authored-by: Jared Norman <jared@super.gd>