Allow using the variant_search_class in API
Our admin has already mostly moved over to using the new search class in
a few places. This adds it as an option to the API index endpoint.

After this change, we'll be able to update our autocomplete JS logic to
use the new search object instead of the hardcoded ransack queries it
currently does. Eventually allowing us a much easier configuration
interface we can use to ensure the variant searching works for all users
who wan't to tweak it to include other search fields. (e.g.: Globalize
wants to search different translated columns.)

The existing behaviour is also easy enough to keep in, which prevents us
from having to make a breaking change.