Fix AddPaymentSourcesToWallet changing default when reused
PR #2913 [1] changed how the new default WalletPaymentSource was set.
Before:
- using the last one found/created by the Order's PaymentSources

After:
- using the user's last WalletPaymentSource.

This is problematic when an order uses a default WalletPaymentSource
but it's not the most recent one.

Fixes issue #4004 [2]

Now, #make_default uses WalletPaymentSources found or created by the
order, which means it will keep the default payment source if used.

[1] https://github.com/solidusio/solidus/pull/2913/
[2] https://github.com/solidusio/solidus/issues/4004