Fix ReDos vulnerability on Spree::EmailValidator::EMAIL_REGEXP
The denial of service vulnerability could be exploited in the checkout
process for guest orders, as it's the only place where that regexp is
used. To this point, registered users' emails (for instance, through
solidus_auth_devise) don't validate against that regular expression.

The previous regular expression had `([^@\s]+\.)+` bit on it. It was
susceptible to exponential backtracking with a substring like `a.a.`.
I.e.:

```
irb(main)> Spree::EmailValidator::EMAIL_REGEXP.match "a@a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.a.@"
processing time: 54.293660s
=> nil
```

We take the occasion to remove the crafted regexp in use altogether.
Instead, we use `URI::MailTo::REGEXP`, which follows
https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address.

It's improbable to find old emails that become invalid in a system,
but, as a safeguard, we provide a task
`solidus:check_orders_with_invalid_emails` to print information for
orders that don't fulfill the new requirement.

The issue has been present since 7af4b932. I.e., since version 1.0. The
regexp had been copied from Devise's one, but they updated it in
https://github.com/heartcombo/devise/commit/830d3e86ee9541b1f8a9ddddd176bba35e0a0e68

References:

https://github.com/solidusio/solidus/security/advisories/GHSA-qxmr-qxh6-2cc9
https://en.wikipedia.org/wiki/ReDoS
https://snyk.io/blog/redos-and-catastrophic-backtracking/

References #4022