Normalize promotion codes before validation
Promotion codes are always stored in the database in lowercase. Up to
this commit, the transformation of the in-memory value of a code was
performed in a `before_save` callback.

This implementation led to a weird behavior that can easily be
reproduced with a fresh Solidus install:
- create a first promotion, with a single promotion code (e.g. "test")
- attempt to create another promotion, with the exact same value for the
  single promotion code
    => triggers a flash error messages which states "Codes is invalid"
- attempt to re-save the same form, this time with a single promotion
  code with the value "TEST"
    => the page crashes due to an `ActiveRecord::RecordNotUnique` error

The expected error handling would be that any input that would end up
matching an existing code (after having been stripped of spaces &
downcased) should trigger the same flash error rather than crash the
page.

This behavior can be achieved by normalizing values earlier in a
`before_validation` callback: that way, the validation callback will
query the database using a stripped/downcased value and, should a match
be found, set an error on the model instance as expected.