Check for edit ability when showing edit links
Instead of checking for the ability to `update`, we should check for
the `edit` ability when showing edit links instead. This is because
the edit page often works also as a replacement for show, and the
list page may not include all the details for a given resource.

On the edit page, the `update` ability will discriminate if the user
can finalize changes or not.