Add image preference spec test
Introduce tests to verify the configuration options added in
b1fa1b897fc9eed74b2c5fe77f9b621c6e7daa2e.

After trying many different options to have `Spree::Image` re-evaluate
the newly set preferences dynamically, I found this working solution. As
`Spree::Image` is used in other spec tests it might already be loaded,
and therefore has already evaluated the previous preferences.

In this solution I decided to make use of a test-specific class that is
equivalent to `Spree::Image`, but uses a fixed include (instead of a
dynamic one).  Each test contains one of the test-specific classes with
a fixed include that references one of the potential
`image_attachment_module`s.

By including the custom class in the test we get the ability to directly
influence the `Spree::Config` values that will be used upon definition
of the class.  Using the `stub_spree_preferences` we can set these
preferences, without influencing other tests in test suite.

The downside of this approach is that we should reflect changes to the
`Spree::Image` class in the custom classes within (if the changes affect
related behavior).