Add Spree.Views.Order.Address spec tests
Implements two spec tests that simulate both states for the preference
`use_combined_first_and_last_name_in_address` verifying that the name
fields are updated correctly.

The fixtures are simplified versions of the partial
'spree/admin/shared/address_form' as I was unable to use the partial
rendering.  This means that the tests only verify that the
Spree.Views.Order.Address Backbone View correctly tries to update the
fields.